fix: merge conflict
merge 할 때 발생한 충돌 처리
- 그 외 나오는 경고들 수정

+++ TL2/StaticMesh.cpp
@@ -68,7 +68,7 @@ void UStaticMesh::BuildMeshBVH()
     }
 
     TArray<FNarrowPhaseBVHPrimitive> Primitives;
-    const int32 TriangleCount = StaticMeshAsset->Indices.size() / 3;
+    const int32 TriangleCount = static_cast<int32>(StaticMeshAsset->Indices.size()) / 3;
     Primitives.Reserve(TriangleCount);
 
     for (int32 i = 0;i < TriangleCount;++i)
